- moved NoteModify resource into own module
